+++ src/main/java/racingcar/controller/RacingController.java
@@ -1,0 +1,5 @@
+package racingcar.controller;
+
+public class RacingController {
+
+}
