+++ src/main/java/racingcar/domain/CarAction.java
@@ -1,4 +1,4 @@
-package racingcar.car;
+package racingcar.domain;
 
 public enum CarAction {
     STAY, FORWARD
