+++ src/main/java/racingcar/domain/MovingStrategy.java
@@ -1,4 +1,4 @@
-package racingcar.car;
+package racingcar.domain;
 
 @FunctionalInterface
 public interface MovingStrategy {
