+++ src/test/java/racingcar/domain/RacingTest.java
@@ -1,4 +1,4 @@
-package racingcar.racing;
+package racingcar.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
