+++ src/test/java/racingcar/util/ValidatorTest.java
@@ -1,4 +1,4 @@
-package racingcar.racing;
+package racingcar.util;
 
 import static org.assertj.core.api.Assertions.assertThat;
